+++ utils/normalize_unix.go
@@ -1,0 +1,9 @@
+//go:build !windows && !darwin
+
+package utils
+
+import "path/filepath"
+
+func NormalizePath(p string) (string, error) {
+	return filepath.Clean(p), nil
+}
